smart-devices/requirements.txt
prometheus-client==0.9.0
python_kasa==0.4.0.dev2
apscheduler==3.6.3
aiolifx==0.6.8